pyproject.toml
[project]
name = "genie-ai"
version = "0.1.0"
description = "Multi-Agent Research & Reporting Engine (MARRE)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # LangGraph / LangChain
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    # Vector store
    "chromadb>=0.5.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    # LLM clients
    "openai>=1.30.0",
    "anthropic>=0.28.0",
    # Document loaders
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    # Utils
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
